refactor(auth): moves generate query params function to utils
