+++ src/modules/auth/hooks/useAuth.ts
@@ -1,12 +1,16 @@
 import { useFuel } from '@fuels/react';
 import { BakoSafe } from 'bakosafe';
 import { Provider } from 'fuels';
-import { useCallback, useState } from 'react';
+import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 
 import { queryClient } from '@/config';
 
-import { useAuthCookies, useQueryParams } from '..';
+import {
+  generateRedirectQueryParams,
+  useAuthCookies,
+  useQueryParams,
+} from '..';
 import { AuthenticateParams, IUseAuthReturn, TypeUser } from '../services';
 import { useUserInfoRequest } from './useUserInfoRequest';
 
@@ -32,24 +36,16 @@ const useAuth = (): IUseAuthReturn => {
     setAuthCookies(params);
   };
 
-  const generateLogoutURLQueryParams = useCallback(() => {
-    const queryParams = [
-      sessionId && `sessionId=${sessionId}`,
-      origin && `origin=${origin}`,
-      name && `name=${name}`,
-      request_id && `request_id=${request_id}`,
-    ]
-      .filter(Boolean)
-      .join('&');
-
-    return queryParams ? `?${queryParams}` : '';
-  }, [sessionId, origin, name, request_id]);
-
   const logout = () => {
     clearAuthCookies();
     queryClient.clear();
 
-    const queryParams = generateLogoutURLQueryParams();
+    const queryParams = generateRedirectQueryParams({
+      sessionId,
+      origin,
+      name,
+      request_id,
+    });
     navigate(`/${queryParams}`);
   };
