+++ src/modules/auth/utils.ts
@@ -1,5 +1,12 @@
 import { CookieName, CookiesConfig } from '@/config/cookies';
 
+interface IGenerateRedirectQueryParams {
+  sessionId: string | null;
+  origin?: string | null;
+  name?: string | null;
+  request_id?: string | null;
+}
+
 const authCredentials = () => ({
   token: CookiesConfig.getCookie(CookieName.ACCESS_TOKEN),
   address: CookiesConfig.getCookie(CookieName.ADDRESS),
@@ -16,4 +23,22 @@ const authCredentialsByKey = (key: 'token' | 'address') => {
   return credentials[key];
 };
 
-export { authCredentials, authCredentialsByKey };
+const generateRedirectQueryParams = ({
+  sessionId,
+  origin,
+  name,
+  request_id,
+}: IGenerateRedirectQueryParams) => {
+  const queryParams = [
+    sessionId && `sessionId=${sessionId}`,
+    origin && `origin=${origin}`,
+    name && `name=${name}`,
+    request_id && `request_id=${request_id}`,
+  ]
+    .filter(Boolean)
+    .join('&');
+
+  return queryParams ? `?${queryParams}` : '';
+};
+
+export { authCredentials, authCredentialsByKey, generateRedirectQueryParams };
